PhishingLogicApp/PhishingLA_Sentinel_Comments/FunctionApp/requirements.txt
azure-functions==1.21.3
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
filelock==3.16.1
idna==3.10
pdfminer.six==20240706
pycparser==2.22
requests==2.32.3
requests-file==2.1.0
soupsieve==2.6
tldextract==5.1.3
urllib3==2.2.3
jinja2